MOD00-DesafiosLG/desafio-7-automacao-de-download-de-dados-covid-19-e-criacao-de-grafico-com-python/DataBot-19/requirements.txt
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
botcity-framework-base==0.4.4
botcity-framework-web==0.9.0
botcity-googlesheets-plugin==0.3.0
botcity-maestro-sdk==0.5.1
boto3==1.35.52
botocore==1.35.52
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.5.8
decorator==5.1.1
distro==1.9.0
dynamic-yaml==1.3.5
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.54.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
marshmallow==3.23.0
marshmallow-enum==1.5.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
msedge-selenium-tools==3.141.3
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.5
owid-catalog==0.3.11
owid-repack==0.1.3
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
rdata==0.9
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.20.1
rsa==4.9
s3transfer==0.10.3
selenium==3.141.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
structlog==24.4.0
tomli==2.0.2
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
webdriver-manager==4.0.2
xarray==2024.10.0
zipp==3.20.2
